added arm_math park clarke transform

+++ src/tasks/task_cli.c
@@ -46,7 +46,7 @@ static void cli_task(void * parameters) {
 
 
     // Print Shell Header
-    for (int i = 0; i < strlen(shell_head); i++) {
+    for (size_t i = 0; i < strlen(shell_head); i++) {
         cli_uart_putc(shell_head[i]);
         vTaskDelay(1);
     }
